+++ packages/core/src/node-execution-context/__tests__/execute-context.test.ts
@@ -12,15 +12,11 @@ import type {
 	Expression,
 	INodeType,
 	INodeTypes,
-	OnError,
-	ContextType,
-	IContextObject,
 	ICredentialDataDecryptedObject,
-	ISourceData,
-	ITaskMetadata,
 } from 'n8n-workflow';
-import { ApplicationError, ExpressionError, NodeHelpers } from 'n8n-workflow';
+import { ApplicationError, ExpressionError } from 'n8n-workflow';
 
+import { describeCommonTests } from './shared-tests';
 import { ExecuteContext } from '../execute-context';
 
 describe('ExecuteContext', () => {
@@ -45,6 +41,7 @@ describe('ExecuteContext', () => {
 	const expression = mock<Expression>();
 	const workflow = mock<Workflow>({ expression, nodeTypes });
 	const node = mock<INode>({
+		name: 'Test Node',
 		credentials: {
 			[testCredentialType]: {
 				id: 'testCredentialId',
@@ -59,7 +56,7 @@ describe('ExecuteContext', () => {
 	const additionalData = mock<IWorkflowExecuteAdditionalData>({ credentialsHelper });
 	const mode: WorkflowExecuteMode = 'manual';
 	const runExecutionData = mock<IRunExecutionData>();
-	const connectionInputData = mock<INodeExecutionData[]>();
+	const connectionInputData: INodeExecutionData[] = [];
 	const inputData: ITaskDataConnections = { main: [[{ json: { test: 'data' } }]] };
 	const executeData = mock<IExecuteData>();
 	const runIndex = 0;
@@ -85,79 +82,12 @@ describe('ExecuteContext', () => {
 		expression.getParameterValue.mockImplementation((value) => value);
 	});
 
-	describe('getExecutionCancelSignal', () => {
-		it('should return the abort signal', () => {
-			expect(executeContext.getExecutionCancelSignal()).toBe(abortSignal);
-		});
-	});
-
-	describe('continueOnFail', () => {
-		afterEach(() => {
-			node.onError = undefined;
-			node.continueOnFail = false;
-		});
-
-		it('should return false for nodes by default', () => {
-			expect(executeContext.continueOnFail()).toEqual(false);
-		});
-
-		it('should return true if node has continueOnFail set to true', () => {
-			node.continueOnFail = true;
-			expect(executeContext.continueOnFail()).toEqual(true);
-		});
-
-		test.each([
-			['continueRegularOutput', true],
-			['continueErrorOutput', true],
-			['stopWorkflow', false],
-		])('if node has onError set to %s, it should return %s', (onError, expected) => {
-			node.onError = onError as OnError;
-			expect(executeContext.continueOnFail()).toEqual(expected);
-		});
-	});
-
-	describe('evaluateExpression', () => {
-		it('should evaluate the expression correctly', () => {
-			const expression = '$json.test';
-			const expectedResult = 'data';
-			const resolveSimpleParameterValueSpy = jest.spyOn(
-				workflow.expression,
-				'resolveSimpleParameterValue',
-			);
-			resolveSimpleParameterValueSpy.mockReturnValue(expectedResult);
-
-			expect(executeContext.evaluateExpression(expression)).toEqual(expectedResult);
-
-			expect(resolveSimpleParameterValueSpy).toHaveBeenCalledWith(
-				`=${expression}`,
-				{},
-				runExecutionData,
-				runIndex,
-				0,
-				node.name,
-				connectionInputData,
-				mode,
-				expect.objectContaining({}),
-				executeData,
-			);
-
-			resolveSimpleParameterValueSpy.mockRestore();
-		});
-	});
-
-	describe('getContext', () => {
-		it('should return the context object', () => {
-			const contextType: ContextType = 'node';
-			const expectedContext = mock<IContextObject>();
-			const getContextSpy = jest.spyOn(NodeHelpers, 'getContext');
-			getContextSpy.mockReturnValue(expectedContext);
-
-			expect(executeContext.getContext(contextType)).toEqual(expectedContext);
-
-			expect(getContextSpy).toHaveBeenCalledWith(runExecutionData, contextType, node);
-
-			getContextSpy.mockRestore();
-		});
+	describeCommonTests(executeContext, {
+		abortSignal,
+		node,
+		workflow,
+		executeData,
+		runExecutionData,
 	});
 
 	describe('getInputData', () => {
@@ -277,54 +207,6 @@ describe('ExecuteContext', () => {
 				'last',
 				'params',
 			]);
-		});
-	});
-
-	describe('getInputSourceData', () => {
-		it('should return the input source data correctly', () => {
-			const inputSourceData = mock<ISourceData>();
-			executeData.source = { main: [inputSourceData] };
-
-			expect(executeContext.getInputSourceData()).toEqual(inputSourceData);
-		});
-
-		it('should throw an error if the source data is missing', () => {
-			executeData.source = null;
-
-			expect(() => executeContext.getInputSourceData()).toThrow(ApplicationError);
-		});
-	});
-
-	describe('logAiEvent', () => {
-		it('should log the AI event correctly', () => {
-			const eventName = 'ai-tool-called';
-			const msg = 'test message';
-
-			executeContext.logAiEvent(eventName, msg);
-
-			expect(additionalData.logAiEvent).toHaveBeenCalledWith(eventName, {
-				executionId: additionalData.executionId,
-				nodeName: node.name,
-				workflowName: workflow.name,
-				nodeType: node.type,
-				workflowId: workflow.id,
-				msg,
-			});
-		});
-	});
-
-	describe('setMetadata', () => {
-		it('sets metadata on execution data', () => {
-			const metadata: ITaskMetadata = {
-				subExecution: {
-					workflowId: '123',
-					executionId: 'xyz',
-				},
-			};
-
-			expect(executeContext.getExecuteData().metadata?.subExecution).toEqual(undefined);
-			executeContext.setMetadata(metadata);
-			expect(executeContext.getExecuteData().metadata?.subExecution).toEqual(metadata.subExecution);
 		});
 	});
 });
