+++ packages/core/src/node-execution-context/__tests__/supply-data-context.test.ts
@@ -12,11 +12,11 @@ import type {
 	Expression,
 	INodeType,
 	INodeTypes,
-	OnError,
 	ICredentialDataDecryptedObject,
 } from 'n8n-workflow';
 import { ApplicationError } from 'n8n-workflow';
 
+import { describeCommonTests } from './shared-tests';
 import { SupplyDataContext } from '../supply-data-context';
 
 describe('SupplyDataContext', () => {
@@ -41,6 +41,7 @@ describe('SupplyDataContext', () => {
 	const expression = mock<Expression>();
 	const workflow = mock<Workflow>({ expression, nodeTypes });
 	const node = mock<INode>({
+		name: 'Test Node',
 		credentials: {
 			[testCredentialType]: {
 				id: 'testCredentialId',
@@ -54,7 +55,7 @@ describe('SupplyDataContext', () => {
 	const additionalData = mock<IWorkflowExecuteAdditionalData>({ credentialsHelper });
 	const mode: WorkflowExecuteMode = 'manual';
 	const runExecutionData = mock<IRunExecutionData>();
-	const connectionInputData = mock<INodeExecutionData[]>();
+	const connectionInputData: INodeExecutionData[] = [];
 	const inputData: ITaskDataConnections = { main: [[{ json: { test: 'data' } }]] };
 	const executeData = mock<IExecuteData>();
 	const runIndex = 0;
@@ -80,64 +81,12 @@ describe('SupplyDataContext', () => {
 		expression.getParameterValue.mockImplementation((value) => value);
 	});
 
-	describe('getExecutionCancelSignal', () => {
-		it('should return the abort signal', () => {
-			expect(supplyDataContext.getExecutionCancelSignal()).toBe(abortSignal);
-		});
-	});
-
-	describe('continueOnFail', () => {
-		afterEach(() => {
-			node.onError = undefined;
-			node.continueOnFail = false;
-		});
-
-		it('should return false for nodes by default', () => {
-			expect(supplyDataContext.continueOnFail()).toEqual(false);
-		});
-
-		it('should return true if node has continueOnFail set to true', () => {
-			node.continueOnFail = true;
-			expect(supplyDataContext.continueOnFail()).toEqual(true);
-		});
-
-		test.each([
-			['continueRegularOutput', true],
-			['continueErrorOutput', true],
-			['stopWorkflow', false],
-		])('if node has onError set to %s, it should return %s', (onError, expected) => {
-			node.onError = onError as OnError;
-			expect(supplyDataContext.continueOnFail()).toEqual(expected);
-		});
-	});
-
-	describe('evaluateExpression', () => {
-		it('should evaluate the expression correctly', () => {
-			const expression = '$json.test';
-			const expectedResult = 'data';
-			const resolveSimpleParameterValueSpy = jest.spyOn(
-				workflow.expression,
-				'resolveSimpleParameterValue',
-			);
-			resolveSimpleParameterValueSpy.mockReturnValue(expectedResult);
-
-			expect(supplyDataContext.evaluateExpression(expression, 0)).toEqual(expectedResult);
-
-			expect(resolveSimpleParameterValueSpy).toHaveBeenCalledWith(
-				`=${expression}`,
-				{},
-				runExecutionData,
-				runIndex,
-				0,
-				node.name,
-				connectionInputData,
-				mode,
-				expect.objectContaining({}),
-				executeData,
-			);
-
-			resolveSimpleParameterValueSpy.mockRestore();
-		});
+	describeCommonTests(supplyDataContext, {
+		abortSignal,
+		node,
+		workflow,
+		executeData,
+		runExecutionData,
 	});
 
 	describe('getInputData', () => {
@@ -217,24 +166,6 @@ describe('SupplyDataContext', () => {
 				'last',
 				'params',
 			]);
-		});
-	});
-
-	describe('logAiEvent', () => {
-		it('should log the AI event correctly', () => {
-			const eventName = 'ai-tool-called';
-			const msg = 'test message';
-
-			supplyDataContext.logAiEvent(eventName, msg);
-
-			expect(additionalData.logAiEvent).toHaveBeenCalledWith(eventName, {
-				executionId: additionalData.executionId,
-				nodeName: node.name,
-				workflowName: workflow.name,
-				nodeType: node.type,
-				workflowId: workflow.id,
-				msg,
-			});
 		});
 	});
 });
